Use switch-case instead of chained if's
Doesn't really help performance yet, as the
generated code is almost the same.

+++ src/NewRemoting/RemotingReferenceType.cs
@@ -34,7 +34,7 @@ namespace NewRemoting
 		BoolFalse,
 		Float,
 		Double,
-		Single,
+		Half,
 		String,
 		AddEvent,
 		RemoveEvent,
